+++ src/components/Header.tsx
@@ -3,13 +3,19 @@ import { Button } from './Button';
 import { Link, UseNavigateResult, useNavigate } from '@tanstack/react-router';
 import { Trans } from '@lingui/react/macro';
 import { t } from '@lingui/core/macro';
+import { RouteNameType } from '../core/types/routes-name';
 
 export const Header: React.FC<{ isConnected?: boolean }> = ({ isConnected = false }) => {
   const navigate: UseNavigateResult<string> = useNavigate();
   return (
     <header className="header" role="banner">
       <div className="header__inner">
-        <Link to="/" className="header__logo-container" aria-label={t`Freelance home page`} viewTransition>
+        <Link
+          to={RouteNameType.Home.toString()}
+          className="header__logo-container"
+          aria-label={t`Freelance home page`}
+          viewTransition
+        >
           <span className="material-symbols-outlined header__logo-icon">terminal</span>
           <span className="header__logo-text">Freelance Dev</span>
         </Link>
@@ -18,7 +24,7 @@ export const Header: React.FC<{ isConnected?: boolean }> = ({ isConnected = fals
           <ul className="header__nav-list">
             <li>
               <Link
-                to="/projects"
+                to={RouteNameType.Projects.toString()}
                 className="header__nav-link"
                 activeProps={{
                   className: 'header__nav-link--active',
@@ -30,7 +36,7 @@ export const Header: React.FC<{ isConnected?: boolean }> = ({ isConnected = fals
             </li>
             <li>
               <Link
-                to="/about"
+                to={RouteNameType.AboutMe.toString()}
                 className="header__nav-link"
                 activeProps={{ className: 'header__nav-link--active' }}
                 viewTransition
@@ -42,7 +48,7 @@ export const Header: React.FC<{ isConnected?: boolean }> = ({ isConnected = fals
             {/* Route disabled until the backend is ready
             <li>
               <Link
-                to="/services"
+                to={RouteNameType.Services.toString()}
                 className="header__nav-link"
                 activeProps={{ className: 'header__nav-link--active' }}
                 viewTransition
@@ -58,14 +64,14 @@ export const Header: React.FC<{ isConnected?: boolean }> = ({ isConnected = fals
             <Button
               className="btn btn--secondary  hero__btn"
               name={t`Admin Dashboard`}
-              onClick={() => navigate({ to: '/admin/dashboard' })}
+              onClick={() => navigate({ to: RouteNameType.AdminDashboard.toString() })}
             >
               <span className="material-symbols-outlined">admin_panel_settings</span>
             </Button>
           )}
           {/* {!isConnected && (
             <Link
-              to="/login"
+              to={RouteNameType.Login.toString()}
               className="header__nav-link"
               activeProps={{ className: 'header__nav-link--active' }}
               viewTransition