+++ src/features/theme/utils/theme.types.ts
@@ -1,1 +1,8 @@
-export type ThemeType = 'nature' | 'dark';
+export type ThemeType = 'nature' | 'dark' | 'light' | 'ocean' | 'desert';
+
+export interface ITheme {
+  id: ThemeType;
+  name: string;
+  primary: string;
+  secondary: string;
+}